Added New toolbar code as Partial in a separate file
