+++ Editor/SceneViewToolbar/SceneViewToolbar.Overlays.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c55c9d2a68d570945b9dccea2290b061
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
